requirements-dev.txt
-r requirements.txt
WebTest==2.0.16
django-webtest==1.7.7
factory-boy==2.4.1
mock==1.0.1
pytest-django==2.8.0
pytest-mock==0.4.0
pytest-xdist==1.11
pytest==2.6.4
tox==1.9.0
